reports edits

+++ backend/src/CCE.Application/Reports/Queries/GetCommunityPostReport/GetCommunityPostReportQueryHandler.cs
@@ -29,7 +29,7 @@ internal sealed class GetCommunityPostReportQueryHandler(
                 p.Id,
                 p.Title,
                 p.Content,
-                (int)p.Type,
+                p.Type,
                 p.CreatedOn),
             q.Page,
             q.PageSize,
